feat: add new token for disabled-text #195
This commit creates a new semantic token specifically to be used in UIs
where the content of an interaction is to appear as disabled to the end
user. This new token may be used in cases where there is unavailable
features in the UI that are not specifically interactive elements.

On branch dsande/disabledText/195
Changes to be committed:
modified:   src/color/text.json